Renamed the new classes with generic key types. Generated the classes with "double" key type with same names as were used before.
