+++ src/FibonacciHeap/FibonacciHeap.csproj
@@ -41,8 +41,10 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="FibonacciHeap.cs" />
+    <Compile Include="FibonacciHeapGenericKey.cs" />
+    <Compile Include="FibonacciHeapNodeGenericKey.cs" />
     <Compile Include="FibonacciHeapNode.cs" />
+    <Compile Include="FibonacciHeap.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
